Overhauling the object code pretty heavily.  Basically stripping it down.  A class now only defines three things: a constructor, a destructor, and a comparator. object_t is no longer a pointer. ctor/dtor/comparator function pointers have their own type aliases now. superClass is now just super.  That's more of a convenience change.
Reason for these changes:
This was originally going to be modeled after Obj-C and whatnot, doing messaging and so on.  Doesn't really work out all that well when you think about it long enough - Obj-C works because of its syntax, but doing this in straight C is not very pleasant.  So, I scaled it back to where I think this is still of use: you can allocate an object of any type by passing a class and a memory pool, anything that follows the same object design can be retained, released, autoreleased, and compared (i.e., for map keys and such).  All objects must have a constructor/destructor, which while not useful for everything gives every object a chance to initialize its members and later properly clean up after itself.