Changing object allocation.
Some changes had to be made to the design of the whole object
allocation system and whatnot - it wasn't holding up all that well.
So, I decided to decouple allocation from initialization as in
Objective-C.  Objects still have constructors, but that's more for the
purpose of sanitizing them.  They are technically usable after
construction, though probably not in the most ideal state.

So, all objects are now allocated via an object_new (except when the
class provides its own _new function [see entity_new, which provides
entity_new to ensure entities are allocated from the entity memory
pool]) and <classname>_init pair (i.e.,
array_init(object_new(array_class, pool), ...)).  This may be
simplified later with a special macro, but I'm not sure about that
yet.

Memory pools are now passed to constructors (why I didn't do this
before is beyond me, but it's very useful for the object to have
access to the pool it belongs to on construction, just in case it
needs to do its own allocations to prepare).  As usual, these pools
can be NULL in order to refer to the global memory pool.

I'll have to write up some style guidelines for this, just so it's
clear what's going on here.