Adding a convenience method for creating objects.
Basically, pass a class *name* (not the identifier for the actual class)
and a pool and any initialization arguments to obj_new and it'll wrap a
call to object_new in an initializer for that class.  This assumes that
all classes have initialization functions - which they probably will
after I mandate it for the sake of consistency.