Moving all camelcase names to underscores.
The great underscore-ening.  So, the reason for this: identifiers in
snow have, thus far, been a pretty odd mishmash of underscores and camel
case and so on.  This got somewhat annoying, because although I know
what an identifier says, the style was inconsistent.  Suffice to say,
this led to some problems with determining what was what.  So, the law
of snow-palm is now pretty simple: underscores or death.  This will
infuriate a few people, but I'm basing this off the C standard library.
I still need to write up style guidelines for this stuff.