Expanding the config file.
Added Mac, iOS, the iOS Simulator, QNX, and Android to the list of
platform macros.  Lined up their indentation as well.

Lined up indentation for architectures.

Gave __USE_UNIX98 a value of 1 just to be explicit.

Gave USE_PTHREADS a concrete value instead of a logical expression.

Added includes for maths/ types to the end of the config.