Removing unused propagate member from event_t.
I'm removing propagate because it's unused. Essentially, the return
value of a handler (a boolean) offers the same information. If the
event should cease propagation, the handler returns true. This does not
mean the handler has done anything about the event (it may be filtering
events out of the system), but it does mean the event will reach no
more handlers. So, a propagate member in an event is a bit useless.