Moving array_t to the init model (e.g., list_t).
This meant renaming array_new to array_init and giving it a self arg.

array_destroy now zeroes the array structure instead of deallocating
it. This is mainly because it can't guarantee the array was allocated
at all, much less whether it was allocated using a given allocator.
This isn't the reason for moving to the init model, just why I think
zeroing works best.

array_copy has been modified accordingly, taking both input and result
arguments.Also went through a few changes to reduce allocations:
assuming the destination is a previously initialized array, which it
should be, it will be checked to see if it has a buffer and, if it
does, whether the existing buffer can hold the source array. If it can,
it will reuse the buffer (after zeroing it -- which happens even if it
can't reuse the buffer).