Moving serializer to stream API.
This was simpler than I expected. Added sz_read_with_bail macro to cut
down on a bit of code repetition. Also added a few more read checks,
so that should fill it out.

So, now the serializer doesn't depend on files (though it still depends
on buffers -- not concerned about that).