build-sources: Add check for 'exclude' file.
If the exclude file is present, that directory and all of its files and
subdirectories are ignored in generating the Makefile. This is mostly
for convenience when testing because I have a lot of broken source
sitting around and it prevents me from doing tests on other things that
aren't heavily work-in-progress.