streams: Replace stdio with PhysFS for streams.
This required a change: STREAM_READWRITE is no longer usable.

STREAM_APPEND replaced it.  This is intended to function the same as
'a' mode streams (going by fopen) or O_WRONLY|O_APPEND (open). Makes
it less flexible, but overall it's probably not going to be missed --
the chance that I'll need to read and write at the same time seems low.