serializer: Use stream primitive functions.
This gives them the benefit of endianness conversion without building
endianness into the serializer itself.

Additionally, sz_read/write_bytes now take a void** or void* for out/in
arguments. This avoids a few needless casts, at least in the case of
sz_write_bytes.

sz_file_error now handles file errors for both read/write operations.
Its name is a bit of a misnomer given the move to streams, but it
should be fine.