src-gen: Add C file definitions to define SNOW_SOURCE.
This is an attempt at trying to work out a way to make inline functions
in headers work a little better. Previously, if an inline was included,
it got an implementation regardless of what happened. So, the trick
for now is to define SNOW_SOURCE only if the header is being included
from its implementation. It then defines SNOW_SOURCE, which snow-config
will use to define S_INLINE as nothing, allowing that source to act as
the implementation for the inline functions if they don't get inlined.
In other source, they should remain inline provided nothing really damn
weird happens.